Added exceptions for keyboardinterrupts
This means that generation of lists can be stopped at any time, and the program will proceed to the next step using already obtained numbers rather than ending abruptly. This could be useful if you accidentally made the range 0-1000 but at 250 decided you wanted to stop. This can be done for every list generation stage. Obviously the overall range will change depending on the time at which you stopped.